+++ src/utils/buildPlugins.ts
@@ -1,6 +1,8 @@
 import { Exposed, Plugin, PluginCreator } from '../../typings/rematch'
+import getExposed from './getExposed'
 
-export default <S>(plugins: Array<PluginCreator<S>>, exposed: Exposed<S>) => plugins.reduce((all, { init }) => {
+export default <S>(pluginConfigs: Array<PluginCreator<S>>) => pluginConfigs.reduce((all, { init }) => {
+  const exposed: Exposed<S> = getExposed(pluginConfigs)
   if (init) {
     const plugin: Plugin<S> = init(exposed)
     if (process.env.NODE_ENV !== 'production') {
